Initial commit
Sample Rust/WebAssembly and JavaScript code implementing the
Floyd-Steinberg dithering algorithm.

Consists of a WebAssembly module written in Rust intended to be built
for the wasm32-unknown-unknown target together with an example HTML page
to load the module and to provide an interface to its functionality. The
sample HTML also includes equivalent plain JavaScript functions which
implement the same functionality in order to allow performance
comparisons.

The JS code in the HTML page uses a CORS-Anywhere server
(https://cors-anywhere.herokuapp.com/) in order to load arbitrary images
without tainting the canvas or requiring opaque responses.